* Refactored to create interfaces for the Subset and RangedSubset and their builders.
