+++ easy-type/pom.xml
@@ -30,11 +30,16 @@
       <scope>test</scope>
     </dependency>
 
-<!--    <dependency>-->
-<!--      <groupId>land.artli</groupId>-->
-<!--      <artifactId>easy-type-unicode-code-generator</artifactId>-->
-<!--      <version>${project.version}</version>-->
-<!--    </dependency>-->
+    <dependency>
+      <groupId>com.ibm.icu</groupId>
+      <artifactId>icu4j-charset</artifactId>
+    </dependency>
+
+    <!--    <dependency>-->
+    <!--      <groupId>land.artli</groupId>-->
+    <!--      <artifactId>easy-type-unicode-code-generator</artifactId>-->
+    <!--      <version>${project.version}</version>-->
+    <!--    </dependency>-->
 
   </dependencies>
