+++ easy-type/src/main/java/land/artli/easytype/Subset.java
@@ -2,27 +2,45 @@ package land.artli.easytype;
 
 import java.util.Collections;
 import java.util.List;
+import land.artli.easytype.RangedSubsetImpl.RangedSubsetBuilderImpl;
 
 public interface Subset {
 
-  boolean isInSubset(final int codePoint);
+  boolean isEmpty();
 
-  default boolean isInSubset(final char ch) {
-    return isInSubset((int) ch);
+  default boolean isNotEmpty() {
+    return !isEmpty();
   }
 
-  default boolean isNotInSubset(final int codePoint) {
-    return !isInSubset(codePoint);
+  boolean contains(final int codePoint);
+
+  default boolean contains(final char ch) {
+    return contains((int) ch);
   }
 
-  default boolean isNotInSubset(final char ch) {
-    return !isInSubset((int) ch);
+  default boolean doesNotContain(final int codePoint) {
+    return !contains(codePoint);
   }
 
-  static CompositeSubset of(final Subset... subsets) {
+  default boolean doesNotContain(final char ch) {
+    return !contains((int) ch);
+  }
+
+  /**
+   * <p>Creates a {@link SubsetBuilder} to create {@link Subset} instances.</p>
+   *
+   * <p>Package-scoped because we should only be invoking this as an implementation detail.</p>
+   *
+   * @return A {@link RangedSubsetBuilder} to create {@link RangedSubset} instances.
+   */
+  static SubsetBuilder builder() {
+    return new RangedSubsetBuilderImpl();
+  }
+
+  static Subset of(final Subset... subsets) {
     if (subsets == null || subsets.length == 0) {
-      return Collections::emptyList;
+      return (CompositeSubset) Collections::emptyList;
     }
-    return () -> List.of(subsets);
+    return (CompositeSubset) () -> List.of(subsets);
   }
 }
