+++ easy-type/src/test/java/land/artli/easytype/TypeParser_ConvertCharacterTest.java
@@ -9,12 +9,12 @@ class TypeParser_ConvertCharacterTest extends AbstractTypeParserTest {
 
   @ParameterizedTest
   @CsvSource(value = {
-      "11052|11052",
-      "7IOSZ|11052",
+      "111052|111052",
+      "7ILOSZ|111052",
       "OIZS-ABCD|0125-ABCD",
       "OIZS_ABCD|0125-ABCD",
       "OIZS ABCD|0125-ABCD",
-      "0123-4567-89AB-CDEF-GHIJ-KLMN-OPQR-STUV-WXYZ|0123-4561-89AB-CDEF-GH1J-KLMN-0PQR-5TUV-WXY2",
+      "0123-4567-89AB-CDEF-GHIJ-KLMN-OPQR-STUV-WXYZ|0123-4561-89AB-CDEF-GH1J-K1MN-0PQR-5TUV-WXY2",
   }, delimiter = '|')
   void should_parse_preserving_case(final String value, final String expected) throws ParseException {
 
@@ -25,6 +25,7 @@ class TypeParser_ConvertCharacterTest extends AbstractTypeParserTest {
             .toUpperCase()
             .convertChar('7', '1')
             .convertChar('I', '1')
+            .convertChar('L', '1')
             .convertChar('O', '0')
             .convertChar('S', '5')
             .convertChar('Z', '2')
@@ -34,7 +35,8 @@ class TypeParser_ConvertCharacterTest extends AbstractTypeParserTest {
             .acceptCharRange('0', '6')
             .acceptCharRange('8', '9')
             .acceptCharRange('A', 'H')
-            .acceptCharRange('J', 'N')
+            .acceptCharRange('J', 'K')
+            .acceptCharRange('M', 'N')
             .acceptCharRange('P', 'R')
             .acceptCharRange('T', 'Y')
             .build();
